add chatroom repo
